Use status bar message for lsp initialization
Close #974.
Refactor lsp functions from debugger decorations module into lsp module.
